general refactoring changed model name from session to payload changed hash input from symbol to string
